backend: setting mutations with mongo
