+++ server-api/src/lib/graphql/schema.js
@@ -1,7 +1,11 @@
 const graphql = require('graphql');
 
 const {
- GraphQLObjectType, GraphQLString, GraphQLSchema, GraphQLID 
+  GraphQLObjectType,
+  GraphQLString,
+  GraphQLBoolean,
+  GraphQLSchema,
+  GraphQLID,
 } = graphql;
 
 const dummy = [
@@ -25,6 +29,51 @@ const dummy = [
   },
 ];
 
+const dummy2 = [
+  {
+    id: '1',
+    username: 'abc',
+    name: 'abc',
+    webinars: '2',
+    verified: 'true',
+    password: 'asasvbasfbvasfb',
+    bio: 'nope',
+    location: 'IN',
+    links: 'NA',
+    email: 'abc@gmail.com',
+    followers: '4',
+    expertise: '2',
+  },
+  {
+    id: '2',
+    username: 'def',
+    name: 'def',
+    webinars: '3',
+    verified: 'false',
+    password: 'svasasdvasv',
+    bio: 'nope',
+    location: 'IN',
+    links: 'NA',
+    email: 'def@gmail.com',
+    followers: '2',
+    expertise: '1',
+  },
+  {
+    id: '3',
+    username: 'ghi',
+    name: 'ghi',
+    webinars: '2',
+    verified: 'true',
+    password: 'adsasvasvwasrb',
+    bio: 'nope',
+    location: 'IN',
+    links: 'NA',
+    email: 'ghi@gmail.com',
+    followers: '8',
+    expertise: '9',
+  },
+];
+
 const ConferenceType = new GraphQLObjectType({
   name: 'conference',
   fields: () => ({
@@ -32,6 +81,24 @@ const ConferenceType = new GraphQLObjectType({
     admin: { type: GraphQLString },
     description: { type: GraphQLString },
     link: { type: GraphQLString }, // topics, date and mentees left
+  }),
+});
+
+const MentorType = new GraphQLObjectType({
+  name: 'mentor',
+  fields: () => ({
+    id: { type: GraphQLID },
+    username: { type: GraphQLString },
+    name: { type: GraphQLString },
+    webinars: { type: GraphQLString }, // Type will be changed to `WebinarTypes`
+    verified: { type: GraphQLBoolean },
+    password: { type: GraphQLString },
+    bio: { type: GraphQLString },
+    location: { type: GraphQLString },
+    links: { type: GraphQLString }, // Type will be changed to `{}_public profile links_`
+    email: { type: GraphQLString },
+    followers: { type: GraphQLString }, // Type will be changed to `menteeTypeID`
+    expertise: { type: GraphQLString }, // Type will be changed to `[]_A list of expertise_`
   }),
 });
 
@@ -45,7 +112,17 @@ const rootQuery = new GraphQLObjectType({
       },
       resolve(parent, args) {
         return dummy[0];
-        // this is where we right code to get data from db
+        // this is where we write code to get data from db
+      },
+    },
+    mentor: {
+      type: MentorType,
+      args: {
+        id: { type: GraphQLID },
+      },
+      resolve(parent, args) {
+        return dummy2[0];
+        // this is where we write code to get data from db
       },
     },
   },
